Enable preview pictures for links to team-coder.com
This is done via the Open Graph protocol which is already automatically confirmed to by the Jekyll SEO Tag plugin. But in order to know which picture to use, the plugin needs it specified in the front matter of the article markdown file. The default property name is "image" so I just renamed and used the already existing "title_image" property.

Now in the <head> of the page there is the meta tag for the preview image, something like this:
<meta property="og:image" content="http://team-coder.com/images/posts/2017-11-21-the-problem-with-high-test-coverage/title-image.jpg"/>